+++ appcues/src/main/java/com/appcues/data/remote/response/styling/StyleSizeResponse.kt
@@ -4,6 +4,6 @@ import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 internal data class StyleSizeResponse(
-    val width: Int,
-    val height: Int,
+    val width: Double,
+    val height: Double,
 )
